@@ -1,137 +1,203 @@
 {
   "name": "auth0-delegated-admin",
-  "version": "2.0.0",
+  "version": "3.0.0",
   "description": "Delegated Administration Dashboard",
   "engines": {
-    "node": "6.2.0"
+    "node": "8.9"
   },
   "main": "server.js",
   "scripts": {
-    "clean": "rimraf dist",
-    "client:build": "npm run clean && cross-env NODE_ENV=production webpack --config ./build/webpack/config.prod.js --colors -p",
+    "clean": "rimraf ./dist",
+    "build": "npm run clean && npm run client:build && npm run extension:build",
+    "client:build": "cross-env NODE_ENV=production webpack --config ./build/webpack/config.prod.js --colors -p",
+    "extension:build": "a0-ext build:server ./webtask.js ./dist && cp ./dist/auth0-delegated-admin.extension.$npm_package_version.js ./build/bundle.js",
     "serve:dev": "cross-env NODE_ENV=development nodemon -e js --ignore assets/app/ --ignore build/webpack/ --ignore client/ --ignore server/data.json --ignore node_modules/ ./build/webpack/server.js",
     "serve:prod": "cross-env NODE_ENV=production node index.js",
-    "test": "npm run test:pre && cross-env NODE_ENV=test nyc mocha tests/mocha.js ./tests/**/*.tests.js",
-    "test:watch": "npm run test:pre && cross-env NODE_ENV=test mocha tests/mocha.js ./tests/**/*.tests.js --watch",
+    "test": "cross-env NODE_ENV=test nyc mocha --require ignore-styles tests/mocha.js './tests/**/*.tests.js'",
+    "test:watch": "cross-env NODE_ENV=test mocha --require ignore-styles tests/mocha.js './tests/**/*.tests.js' --watch",
     "test:pre": "npm run test:clean && npm run lint:js",
     "test:clean": "rimraf ./coverage && rimraf ./.nyc_output",
-    "extension:build": "webpack -p --config ./build/extension/webpack.config.js",
-    "extension:size": "cross-env NODE_ENV=production webpack -p --config ./build/extension/webpack.config.js --json > ./build/extension/bundle-size.json && node ./build/extension/bundle-size.js",
-    "win-extension:bundle": "copy dist\\auth0-delegated-admin.extension.%npm_package_version%.js build\\bundle.js /y",
-    "build": "npm run clean && npm run client:build && npm run extension:build",
-    "win-build": "cross-env NODE_ENV=production && npm run clean && npm run client:build && npm run extension:build && npm run win-extension:bundle",
-    "lint:js": "eslint --ignore-path .gitignore --ignore-pattern build --ignore-pattern webpack ."
+    "extension:size": "cross-env NODE_ENV=production webpack -p --config ./build/extension/webpack.config.js --json > ./build/extension/bundle-size.json && node ./build/extension/bundle-size.js"
   },
   "keywords": [
     "auth0"
   ],
+  "auth0-extension": {
+    "externals": [
+      "async@2.1.2",
+      "auth0@2.8.0",
+      "aws-sdk@2.5.3",
+      "blipp@2.3.0",
+      "bluebird@3.4.6",
+      "body-parser@1.12.4",
+      "boom@3.2.2",
+      "cookie-parser@1.3.5",
+      "ejs@2.3.1",
+      "express@4.14.0",
+      "iconv-lite@0.4.10",
+      "inert@4.0.1",
+      "joi@9.0.4",
+      "jsonwebtoken@7.1.9",
+      "jwks-rsa@1.1.1",
+      "lodash@3.10.1",
+      "lru-memoizer@1.10.0",
+      "moment@2.11.2",
+      "morgan@1.5.3",
+      "node-uuid@1.4.3",
+      "mime-db@1.10.0",
+      "qs@3.1.0",
+      "request@2.67.0",
+      "superagent@1.2.0",
+      "tough-cookie@2.2.2",
+      "uuid@2.0.1",
+      "webtask-tools",
+      "winston@1.0.0"
+    ],
+    "settings": {
+      "WARN_DB_SIZE": 409600,
+      "MAX_MULTISELECT_USERS": 5,
+      "MULTISELECT_DEBOUNCE_MS": 250,
+      "PER_PAGE": 10
+    }
+  },
   "author": "Sandrino Di Mattia",
   "license": "MIT",
   "dependencies": {
     "async": "^1.5.2",
-    "auth0": "^2.1.0",
-    "auth0-extension-express-tools": "^0.6.0",
-    "auth0-extension-tools": "^0.14.0",
-    "axios": "^0.13.1",
-    "babel": "^6.5.2",
-    "babel-core": "^6.9.1",
-    "babel-plugin-transform-runtime": "^6.9.0",
-    "babel-polyfill": "^6.9.1",
-    "babel-preset-es2015": "^6.9.0",
-    "babel-preset-es2015-loose": "^7.0.0",
-    "babel-preset-stage-0": "^6.5.0",
-    "babel-register": "^6.9.0",
-    "babel-runtime": "^6.9.2",
+    "auth0": "^2.8.0",
+    "auth0-extension-express-tools": "1.1.8",
+    "auth0-extension-tools": "1.3.3",
+    "auth0-extension-ui": "~1.1.6",
+    "auth0-js": "^9.6.0",
+    "axios": "^0.14.0",
+    "babel": "6.23.0",
+    "babel-core": "6.24.1",
+    "babel-plugin-transform-runtime": "6.23.0",
+    "babel-polyfill": "6.23.0",
+    "babel-preset-es2015": "6.24.1",
+    "babel-preset-es2015-loose": "^8.0.0",
+    "babel-preset-stage-0": "6.24.1",
+    "babel-register": "6.24.1",
+    "babel-runtime": "6.23.0",
     "bluebird": "^3.4.0",
     "body-parser": "^1.15.1",
+    "codemirror": "^5.19.0",
+    "cookie-parser": "^1.4.3",
     "ejs": "^2.4.2",
     "expect": "^1.20.2",
-    "express": "^4.13.4",
+    "express": "^4.16.3",
     "express-jwt": "^3.4.0",
     "glob": "^7.0.3",
+    "history": "^2.1.2",
+    "immutable": "^3.8.1",
     "json-loader": "^0.5.4",
     "jsonwebtoken": "^7.1.7",
     "jwks-rsa": "^1.1.0",
+    "jwt-decode": "^2.0.1",
     "lodash": "^3.10.1",
-    "lru-memoizer": "^1.5.0",
+    "lru-memoizer": "1.10.0",
     "moment": "^2.13.0",
     "morgan": "^1.7.0",
     "ms": "^0.7.1",
     "nconf": "^0.8.4",
+    "prop-types": "^15.6.0",
+    "qs": "^6.5.2",
+    "react": "15.6.1",
+    "react-codemirror": "^0.2.6",
+    "react-dom": "15.6.1",
+    "react-redux": "^4.4.5",
+    "react-router": "^2.8.1",
+    "react-router-redux": "^4.0.6",
+    "redux": "^3.5.2",
+    "redux-form": "^6.0.5",
+    "redux-logger": "^2.7.0",
+    "redux-promise-middleware": "^4.1.0",
+    "redux-thunk": "2.2.0",
     "request": "^2.34.0",
+    "reselect": "^2.5.4",
     "safe-eval": "^0.3.0",
     "semver": "^5.1.0",
-    "supertest": "^2.0.0",
-    "webtask-tools": "^2.2.0",
+    "string.prototype.endswith": "^0.2.0",
+    "superagent": "^3.8.3",
+    "uuid": "2.0.1",
+    "webtask-tools": "^3.1.0",
     "winston": "^2.2.0"
   },
   "devDependencies": {
-    "autoprefixer": "^6.3.6",
-    "babel-eslint": "^6.0.4",
-    "babel-loader": "^6.2.4",
-    "babel-preset-react": "^6.5.0",
+    "auth0-extensions-cli": "~1.0.9",
+    "autoprefixer": "^6.5.1",
+    "babel-eslint": "7.2.3",
+    "babel-loader": "7.0.0",
+    "babel-preset-react": "6.24.1",
     "babel-preset-react-hmre": "^1.1.1",
     "chai": "^3.5.0",
+    "chai-match": "^1.1.1",
     "classnames": "^2.2.5",
-    "cross-env": "^2.0.1",
-    "css-loader": "^0.23.1",
-    "eslint": "^3.2.2",
-    "eslint-config-airbnb": "^9.0.1",
-    "eslint-plugin-babel": "^3.2.0",
-    "eslint-plugin-import": "^1.12.0",
-    "eslint-plugin-jsx-a11y": "^2.0.1",
-    "eslint-plugin-react": "^5.2.2",
+    "cross-env": "5.0.0",
+    "css-loader": "0.28.2",
+    "enzyme": "^3.3.0",
+    "enzyme-adapter-react-15": "^1.0.5",
+    "eslint": "^3.19.0",
+    "eslint-config-auth0": "10.2.0",
+    "eslint-config-auth0-base": "^12.1.0",
+    "eslint-plugin-babel": "4.1.1",
+    "eslint-plugin-import": "^2.2.0",
+    "eslint-plugin-jsx-a11y": "^4.0.0",
+    "eslint-plugin-react": "7.0.1",
     "expect": "^1.20.2",
-    "exports-loader": "^0.6.3",
+    "exports-loader": "0.6.4",
     "extract-text-webpack-plugin": "^1.0.1",
-    "file-loader": "^0.9.0",
-    "history": "^2.1.2",
-    "immutable": "^3.8.1",
-    "imports-loader": "^0.6.5",
+    "file-loader": "0.11.1",
+    "ignore-styles": "^5.0.1",
+    "imports-loader": "0.7.1",
+    "jsdom": "^11.5.1",
     "json-loader": "^0.5.4",
-    "jwt-decode": "^2.0.1",
-    "mocha": "^3.0.0",
-    "nock": "^8.0.0",
-    "nodemon": "^1.9.2",
-    "nyc": "^8.1.0",
+    "mocha": "3.4.2",
+    "nock": "9.0.13",
+    "nodemon": "^1.11.0",
+    "nyc": "10.3.2",
     "postcss-focus": "^1.0.0",
-    "postcss-loader": "^0.9.1",
+    "postcss-loader": "^1.0.0",
     "postcss-reporter": "^1.3.3",
-    "postcss-simple-vars": "^1.2.0",
+    "postcss-simple-vars": "^3.0.0",
     "raw-loader": "^0.5.1",
-    "react": "^0.14.8",
-    "react-bootstrap": "^0.30.0",
-    "react-codemirror": "^0.2.6",
-    "react-dom": "^0.14.8",
-    "react-loader-advanced": "^1.4.0",
+    "react-bootstrap": "0.31.0",
+    "react-loader-advanced": "1.6.0",
     "react-pure-render": "^1.0.2",
-    "react-redux": "^4.4.5",
-    "react-router": "^2.6.1",
-    "react-router-redux": "^4.0.5",
     "react-select": "^1.0.0-rc.1",
+    "react-test-renderer": "15.6.1",
     "react-transform-hmr": "^1.0.4",
-    "redux": "^3.5.2",
     "redux-devtools": "^3.3.1",
     "redux-devtools-dock-monitor": "^1.1.1",
     "redux-devtools-log-monitor": "^1.0.11",
-    "redux-form": "^4.2.2",
-    "redux-logger": "^2.6.1",
-    "redux-promise-middleware": "^2.4.0",
     "redux-simple-router": "^2.0.4",
-    "redux-static": "^1.0.0",
-    "redux-thunk": "^2.1.0",
+    "redux-static": "1.1.0",
     "request-promise": "^4.1.1",
-    "reselect": "^2.5.3",
-    "rimraf": "^2.5.2",
-    "style-loader": "^0.13.1",
-    "supertest": "^2.0.0",
-    "url-loader": "^0.5.7",
+    "rimraf": "2.6.1",
+    "style-loader": "0.18.1",
+    "supertest": "^2.0.1",
+    "url-loader": "0.5.8",
     "webpack": "^1.13.1",
-    "webpack-dev-middleware": "^1.6.1",
-    "webpack-dev-server": "^1.14.1",
-    "webpack-hot-middleware": "^2.10.0",
-    "webpack-sources": "^0.1.2",
-    "webpack-stats-plugin": "^0.1.1",
-    "webtask-bundle": "^2.1.1"
+    "webpack-dev-middleware": "1.10.2",
+    "webpack-dev-server": "^1.16.2",
+    "webpack-hot-middleware": "2.18.0",
+    "webpack-sources": "0.2.3",
+    "webpack-stats-plugin": "0.1.5",
+    "webtask-bundle": "3.0.1"
+  },
+  "nyc": {
+    "extension": [
+      ".jsx"
+    ],
+    "include": [
+      "client/**/*.js",
+      "client/**/*.jsx",
+      "server/**/*.js"
+    ],
+    "require": [
+      "babel-register"
+    ],
+    "sourceMap": true,
+    "instrument": true
   }
 }
