@@ -1,5 +1,6 @@
 import axios from 'axios';
 import * as constants from '../constants';
+import { getAppSettings } from './auth';
 
 export function fetchScript(name) {
   return {
@@ -21,7 +22,12 @@ export function updateScript(name, script) {
       type: constants.UPDATE_SCRIPT,
       meta: {
         name,
-        script
+        script,
+        onSuccess: () => {
+          if (name === 'settings') {
+            return dispatch(getAppSettings());
+          }
+        }
       },
       payload: {
         promise: axios.post(`/api/scripts/${name}`, { script }, {
