@@ -1,5 +1,7 @@
-import React, { Component, PropTypes } from 'react';
+import React, { Component } from 'react';
+import PropTypes from 'prop-types';
 import { MenuItem, DropdownButton } from 'react-bootstrap';
+import _ from 'lodash';
 
 export default class UserActions extends Component {
   static propTypes = {
@@ -9,17 +11,38 @@ export default class UserActions extends Component {
     changeUsername: PropTypes.func.isRequired,
     databaseConnections: PropTypes.object.isRequired,
     deleteUser: PropTypes.func.isRequired,
+    changeFields: PropTypes.func.isRequired,
     removeMfa: PropTypes.func.isRequired,
     resendVerificationEmail: PropTypes.func.isRequired,
     resetPassword: PropTypes.func.isRequired,
     unblockUser: PropTypes.func.isRequired,
-    user: PropTypes.object.isRequired
+    user: PropTypes.object.isRequired,
+    role: PropTypes.number.isRequired,
+    userFields: PropTypes.array.isRequired,
+    languageDictionary: PropTypes.object
   }
 
-  state = {
-    user: null,
-    loading: false
-  };
+  constructor(props) {
+    super(props);
+
+    if (props.user) {
+      this.state = {
+        user: props.user.toJS(),
+        loading: props.loading
+      };
+
+      if (props.databaseConnections) {
+        this.state.databaseConnections = props.databaseConnections.toJS();
+      }
+    } else {
+      this.state = {
+        user: null,
+        loading: false
+      };
+    }
+
+    this.state.languageDictionary = props.languageDictionary || {};
+  }
 
   componentWillReceiveProps(nextProps) {
     if (nextProps.user) {
@@ -35,17 +58,35 @@ export default class UserActions extends Component {
         databaseConnections: nextProps.databaseConnections.toJS()
       });
     }
-  }
 
-  shouldComponentUpdate(nextProps) {
-    return nextProps.user !== this.props.user || nextProps.databaseConnections !== this.props.databaseConnections;
+    if (nextProps.languageDictionary) {
+      this.setState({
+        languageDictionary: nextProps.languageDictionary
+      });
+    }
   }
 
   getDeleteAction = (user, loading) => (
     <MenuItem disabled={loading || false} onClick={this.deleteUser}>
-      Delete User
+      {this.state.languageDictionary.deleteUserMenuItemText || 'Delete User'}
     </MenuItem>
   )
+
+  getChangeFieldsAction = (user, loading) => {
+    if (!this.props.userFields || !this.props.userFields.length) {
+      return null;
+    }
+
+    /* Only display this if there are editable fields */
+    const fieldsWithEdit = _.filter(this.props.userFields, field => field.edit);
+    if (fieldsWithEdit.length <= 0) return null;
+
+    return (
+      <MenuItem disabled={loading || false} onClick={this.changeFields}>
+        {this.state.languageDictionary.changeFieldsMenuItemText || 'Change Profile'}
+      </MenuItem>
+    );
+  }
 
   getResetPasswordAction = (user, loading) => {
     if (!this.state.databaseConnections || !this.state.databaseConnections.length) {
@@ -54,7 +95,7 @@ export default class UserActions extends Component {
 
     return (
       <MenuItem disabled={loading || false} onClick={this.resetPassword}>
-        Reset Password
+        {this.state.languageDictionary.resetPasswordMenuItemText || 'Reset Password'}
       </MenuItem>
     );
   }
@@ -66,7 +107,7 @@ export default class UserActions extends Component {
 
     return (
       <MenuItem disabled={loading || false} onClick={this.changePassword}>
-        Change Password
+        {this.state.languageDictionary.changePasswordMenuItemText || 'Change Password'}
       </MenuItem>
     );
   }
@@ -76,9 +117,13 @@ export default class UserActions extends Component {
       return null;
     }
 
+    /* Check if settings are disabling the editing of username */
+    const falseUsernameEditFields = _.filter(this.props.userFields, field => field.property === 'username' && field.edit === false);
+    if (falseUsernameEditFields.length > 0) return null;
+
     return (
       <MenuItem disabled={loading || false} onClick={this.changeUsername}>
-        Change Username
+        {this.state.languageDictionary.changeUsernameMenuItemText || 'Change Username'}
       </MenuItem>
     );
   }
@@ -88,9 +133,13 @@ export default class UserActions extends Component {
       return null;
     }
 
+    /* Check if settings are disabling the editing of username */
+    const falseEmailEditFields = _.filter(this.props.userFields, field => field.property === 'email' && field.edit === false);
+    if (falseEmailEditFields.length > 0) return null;
+
     return (
       <MenuItem disabled={loading || false} onClick={this.changeEmail}>
-        Change Email
+        {this.state.languageDictionary.changeEmailMenuItemText || 'Change Email'}
       </MenuItem>
     );
   }
@@ -100,9 +149,13 @@ export default class UserActions extends Component {
       return null;
     }
 
+    /* Check if resending verification email option is enabled */
+    const falseTriggerEmailVerified = _.filter(this.props.userFields, field => field.property === 'email_verified' && field.edit === false);
+    if (falseTriggerEmailVerified.length > 0) return null;
+
     return (
       <MenuItem disabled={loading || false} onClick={this.resendVerificationEmail}>
-        Resend Verification Email
+        {this.state.languageDictionary.resendVerificationEmailMenuItemText || "Resend Verification Email"}
       </MenuItem>
     );
   }
@@ -114,7 +167,7 @@ export default class UserActions extends Component {
 
     return (
       <MenuItem disabled={loading || false} onClick={this.removeMfa}>
-        Remove MFA ({user.multifactor[0]})
+        {this.state.languageDictionary.removeMfaMenuItemText || "Remove MFA"}
       </MenuItem>
     );
   }
@@ -123,20 +176,24 @@ export default class UserActions extends Component {
     if (user.blocked) {
       return (
         <MenuItem disabled={loading || false} onClick={this.unblockUser}>
-          Unblock User
+          {this.state.languageDictionary.unblockUserMenuItemText || "Unblock User"}
         </MenuItem>
       );
     }
 
     return (
       <MenuItem disabled={loading || false} onClick={this.blockUser}>
-        Block User
+        {this.state.languageDictionary.blockUserMenuItemText || "Block User"}
       </MenuItem>
     );
   }
 
   deleteUser = () => {
     this.props.deleteUser(this.state.user);
+  }
+
+  changeFields = () => {
+    this.props.changeFields(this.state.user);
   }
 
   resetPassword = () => {
@@ -147,12 +204,28 @@ export default class UserActions extends Component {
     this.props.changePassword(this.state.user, this.state.databaseConnections[0]);
   }
 
+  static getDisplayObject(user, fields) {
+    if (fields.length > 0) {
+      let displayFunction = undefined;
+      if (_.isFunction(fields[0].edit.display)) displayFunction = fields[0].edit.display;
+      else if (!fields[0].edit.display && fields[0].edit.display !== false && _.isFunction(fields[0].display)) displayFunction = fields[0].display;
+      if (displayFunction) return {
+        display: displayFunction,
+        user
+      };
+    }
+
+    return null;
+  }
+
   changeUsername = () => {
-    this.props.changeUsername(this.state.user, this.state.databaseConnections[0]);
+    const usernameEditFields = _.filter(this.props.userFields, field => field.property === 'username' && field.edit !== false && field.edit);
+    this.props.changeUsername(this.state.user, this.state.databaseConnections[0], UserActions.getDisplayObject(this.state.user, usernameEditFields));
   }
 
   changeEmail = () => {
-    this.props.changeEmail(this.state.user, this.state.databaseConnections[0]);
+    const emailEditFields = _.filter(this.props.userFields, field => field.property === 'email' && field.edit !== false && field.edit);
+    this.props.changeEmail(this.state.user, this.state.databaseConnections[0], UserActions.getDisplayObject(this.state.user, emailEditFields));
   }
 
   resendVerificationEmail = () => {
@@ -168,24 +241,28 @@ export default class UserActions extends Component {
   }
 
   removeMfa = () => {
-    this.props.removeMfa(this.state.user, this.state.user.multifactor[0]);
+    this.props.removeMfa(this.state.user, this.state.user.multifactor);
   }
 
   render() {
-    if (!this.state.user) {
+    if (!this.state.user || this.props.role < 1) {
       return null;
     }
 
+    const languageDictionary = this.props.languageDictionary || {};
+    const buttonTitle = languageDictionary.userActionsButton || 'Actions';
+
     return (
-      <DropdownButton bsStyle="success" title="Actions" id="user-actions">
+      <DropdownButton bsStyle="success" title={buttonTitle} id="user-actions">
         {this.getMultifactorAction(this.state.user, this.state.loading)}
         {this.getBlockedAction(this.state.user, this.state.loading)}
         {this.getResetPasswordAction(this.state.user, this.state.loading)}
-        {this.getChangePasswordAction(this.state.user, this.state.loading)}
-        {this.getDeleteAction(this.state.user, this.state.loading)}
+        {this.getResendEmailVerificationAction(this.state.user, this.state.loading)}
         {this.getChangeUsernameAction(this.state.user, this.state.loading)}
         {this.getChangeEmailAction(this.state.user, this.state.loading)}
-        {this.getResendEmailVerificationAction(this.state.user, this.state.loading)}
+        {this.getChangePasswordAction(this.state.user, this.state.loading)}
+        {this.getChangeFieldsAction(this.state.user, this.state.loading)}
+        {this.getDeleteAction(this.state.user, this.state.loading)}
       </DropdownButton>
     );
   }
