@@ -1,10 +1,9 @@
 import React, { PropTypes, Component } from 'react';
-import './UserInfoField.css';
+import './UserInfoField.styles.css';
 
 export default class UserInfoField extends Component {
   static propTypes = {
-    title: PropTypes.string.isRequired,
-    children: PropTypes.object.isRequired
+    title: PropTypes.string.isRequired
   }
 
   shouldComponentUpdate(nextProps) {
