@@ -1,11 +1,14 @@
 import React, { PropTypes, Component } from 'react';
-import { Error, Json, LoadingPanel } from '../Dashboard';
+import { Error, Json, LoadingPanel } from 'auth0-extension-ui';
+import getErrorMessage from '../../utils/getErrorMessage';
 
 export default class UserProfile extends Component {
   static propTypes = {
     error: PropTypes.string,
     loading: PropTypes.bool.isRequired,
-    user: PropTypes.object.isRequired
+    user: PropTypes.object.isRequired,
+    settings: PropTypes.object.isRequired,
+    languageDictionary: PropTypes.object
   }
 
   shouldComponentUpdate(nextProps) {
@@ -13,10 +16,10 @@ export default class UserProfile extends Component {
   }
 
   render() {
-    const { user, error, loading } = this.props;
+    const { user, error, loading, settings } = this.props;
     return (
       <LoadingPanel show={loading} animationStyle={{ paddingTop: '5px', paddingBottom: '5px' }}>
-        <Error message={error}>
+        <Error title={this.props.languageDictionary.errorTitle} message={getErrorMessage(this.props.languageDictionary, error, settings.errorTranslator)} >
           <Json jsonObject={user.toJS()} />
         </Error>
       </LoadingPanel>
