@@ -1,8 +1,10 @@
-import React, { Component, PropTypes } from 'react';
+import React, { Component } from 'react';
+import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 
 import { logout } from '../actions/auth';
 import { applicationActions, connectionActions, authActions } from '../actions';
+import { LoadingPanel } from 'auth0-extension-ui';
 
 import Header from '../components/Header';
 
@@ -12,36 +14,65 @@ class App extends Component {
     settings: PropTypes.object,
     issuer: PropTypes.string,
     logout: PropTypes.func,
+    settingsLoading: PropTypes.bool,
+    styleSettings: PropTypes.bool,
     fetchApplications: PropTypes.func.isRequired,
     fetchConnections: PropTypes.func.isRequired,
     getAccessLevel: PropTypes.func.isRequired,
-    getAppSettings: PropTypes.func.isRequired
-  }
+    getAppSettings: PropTypes.func.isRequired,
+    toggleStyleSettings: PropTypes.func.isRequired,
+    languageDictionary: PropTypes.object.isRequired
+  };
 
   componentWillMount() {
+    this.props.getAppSettings();
     this.props.fetchApplications();
     this.props.fetchConnections();
     this.props.getAccessLevel();
-    this.props.getAppSettings();
+    this.props.getStyleSettings();
   }
 
   getDictValue = (index, defaultValue) => {
     const appSettings = this.props.settings;
     let val = '';
     if (appSettings.get('settings') && appSettings.get('settings').get('dict')) {
-      val = appSettings.get('settings').get('dict').get(index)
+      val = appSettings.get('settings').get('dict').get(index);
     }
     return val || defaultValue;
-  }
+  };
+
+  onLogout = () => {
+    const appSettings = this.props.settings;
+    let logoutUrl;
+
+    if (appSettings.get('settings') && appSettings.get('settings').get('dict')) {
+      logoutUrl = appSettings.get('settings').get('dict').get('logoutUrl');
+    }
+
+    this.props.logout(logoutUrl);
+  };
 
   render() {
+    const { settingsLoading } = this.props;
+    const languageDictionary = this.props.languageDictionary ? this.props.languageDictionary.toJS() : {};
+    const settings = this.props.settings.get('settings') && this.props.settings.get('settings').toJS();
+    const renderCssToggle = !!(settings && settings.css && settings.altcss);
+
+    if (settingsLoading) {
+      return <LoadingPanel show={settingsLoading} />;
+    }
     return (
       <div>
         <Header
           user={this.props.user}
           issuer={this.props.issuer}
           getDictValue={this.getDictValue}
-          onLogout={this.props.logout} accessLevel={this.props.accessLevel.toJSON()}
+          onLogout={this.onLogout}
+          onCssToggle={this.props.toggleStyleSettings}
+          accessLevel={this.props.accessLevel.toJSON()}
+          styleSettings={this.props.styleSettings}
+          languageDictionary={languageDictionary}
+          renderCssToggle={renderCssToggle}
         />
         <div className="container">
           <div className="row">
@@ -67,9 +98,11 @@ function select(state) {
   return {
     issuer: state.auth.get('issuer'),
     user: state.auth.get('user'),
-    ruleStatus: state.ruleStatus,
     accessLevel: state.accessLevel.get('record'),
-    settings: state.settings.get('record')
+    settings: state.settings.get('record'),
+    styleSettings: state.styleSettings,
+    settingsLoading: state.settings.get('loading'),
+    languageDictionary: state.languageDictionary.get('record')
   };
 }
 
