@@ -1,62 +1,92 @@
-import React, { Component, PropTypes } from 'react';
+import _ from 'lodash';
+import React, { Component } from 'react';
+import PropTypes from 'prop-types';
 import connectContainer from 'redux-static';
+import { Error, Confirm } from 'auth0-extension-ui';
 
+import submitForm from '../../../actions/submitForm';
 import { userActions } from '../../../actions';
-import { Error, Confirm } from '../../../components/Dashboard';
+import { useDisabledConnectionField, useEmailField } from '../../../utils/useDefaultFields';
+import { getName, mapValues } from '../../../utils/display';
+import getDialogMessage from './getDialogMessage';
+import UserFieldsForm from '../../../components/Users/UserFieldsForm';
+import getErrorMessage from '../../../utils/getErrorMessage';
+
 
 export default connectContainer(class extends Component {
   static stateToProps = (state) => ({
-    emailChange: state.emailChange
+    emailChange: state.emailChange,
+    settings: (state.settings.get('record') && state.settings.get('record').toJS().settings) || {},
+    languageDictionary: state.languageDictionary
   });
 
   static actionsToProps = {
+    submitForm,
     ...userActions
-  }
+  };
 
   static propTypes = {
     cancelEmailChange: PropTypes.func.isRequired,
     changeEmail: PropTypes.func.isRequired,
     emailChange: PropTypes.object.isRequired
-  }
+  };
 
   shouldComponentUpdate(nextProps) {
-    return nextProps.emailChange !== this.props.emailChange;
+    return nextProps.languageDictionary !== this.props.languageDictionary ||
+      nextProps.emailChange !== this.props.emailChange;
   }
 
   onConfirm = () => {
-    this.props.changeEmail(this.refs.user.value, this.refs.email.value);
-  }
+    this.props.submitForm('change-email');
+  };
+
+  onSubmit = (emailForm) => {
+    const { user } = this.props.emailChange.toJS();
+
+    this.props.changeEmail(user, emailForm, this.props.languageDictionary.get('record').toJS());
+  };
 
   render() {
-    const { cancelEmailChange } = this.props;
-    const { userId, connection, userEmail, userName, error, requesting, loading } = this.props.emailChange.toJS();
+    const { cancelEmailChange, settings } = this.props;
+    const { user, connection, error, requesting, loading } = this.props.emailChange.toJS();
+
+    const userFields = settings.userFields || [];
+
+    const languageDictionary = this.props.languageDictionary.get('record').toJS();
+
+    const messageFormat = languageDictionary.changeEmailMessage ||
+      'Do you really want to change the email for {username}?';
+    const message = getDialogMessage(messageFormat, 'username',
+      getName(user, userFields, languageDictionary));
+
+    const fields = _.cloneDeep(userFields) || [];
+    useEmailField(true, fields);
+    useDisabledConnectionField(true, fields, connection);
+
+    const allowedFields = ['email', 'connection'];
+    const filteredFields = _.filter(fields,
+      field => _.includes(allowedFields, field.property));
+
+    const UserFieldsFormInstance = UserFieldsForm('change-email', this.onSubmit.bind(this));
 
     return (
       <Confirm
-        title="Change Email?" show={requesting} loading={loading} onCancel={cancelEmailChange}
+        title={languageDictionary.changeEmailTitle || 'Change Email?'}
+        confirmMessage={languageDictionary.dialogConfirmText} cancelMessage={languageDictionary.dialogCancelText}
+        show={requesting} loading={loading} onCancel={cancelEmailChange}
         onConfirm={this.onConfirm}
+        closeLabel={languageDictionary.closeButtonText}
       >
-        <Error message={error} />
+        <Error title={languageDictionary.errorTitle} message={getErrorMessage(languageDictionary, error, settings.errorTranslator)} />
         <p>
-          Do you really want to change the email for <strong>{userName}</strong>?
+          {message}
         </p>
-        <div className="row">
-          <form className="form-horizontal col-xs-12" style={{ marginTop: '40px' }}>
-            <div className="form-group">
-              <label className="col-xs-2 control-label">Connection</label>
-              <div className="col-xs-9">
-                <input type="text" readOnly="readonly" className="form-control" value={connection} />
-              </div>
-            </div>
-            <div className="form-group">
-              <label className="col-xs-2 control-label">Email</label>
-              <div className="col-xs-9">
-                <input ref="email" type="email" className="form-control" defaultValue={userEmail} />
-              </div>
-            </div>
-            <input ref="user" type="hidden" readOnly="readonly" className="form-control" value={userId} />
-          </form>
-        </div>
+        <UserFieldsFormInstance
+          initialValues={mapValues(user, allowedFields, filteredFields, 'edit', languageDictionary)}
+          isEditForm={true}
+          fields={filteredFields}
+          languageDictionary={languageDictionary}
+        />
       </Confirm>
     );
   }
