@@ -1,12 +1,18 @@
-import React, { Component, PropTypes } from 'react';
+import React, { Component } from 'react';
+import PropTypes from 'prop-types';
 import connectContainer from 'redux-static';
+import { Error, Confirm } from 'auth0-extension-ui';
 
 import { userActions } from '../../../actions';
-import { Error, Confirm } from '../../../components/Dashboard';
+import getDialogMessage from './getDialogMessage';
+import { getName } from '../../../utils/display';
+import getErrorMessage from '../../../utils/getErrorMessage';
 
 export default connectContainer(class extends Component {
   static stateToProps = (state) => ({
-    verificationEmail: state.verificationEmail
+    verificationEmail: state.verificationEmail,
+    settings: (state.settings.get('record') && state.settings.get('record').toJS().settings) || {},
+    languageDictionary: state.languageDictionary
   });
 
   static actionsToProps = {
@@ -20,22 +26,40 @@ export default connectContainer(class extends Component {
   }
 
   shouldComponentUpdate(nextProps) {
-    return nextProps.verificationEmail !== this.props.verificationEmail;
+    return nextProps.verificationEmail !== this.props.verificationEmail ||
+      nextProps.languageDictionary !== this.props.languageDictionary;
   }
 
   onConfirm = () => {
-    this.props.resendVerificationEmail(this.props.verificationEmail.toJS().userId);
+    this.props.resendVerificationEmail(this.props.verificationEmail.toJS().user.user_id);
   }
 
   render() {
-    const { cancelResendVerificationEmail } = this.props;
-    const { userName, error, requesting, loading } = this.props.verificationEmail.toJS();
+    const { cancelResendVerificationEmail, settings } = this.props;
+    const { user, error, requesting, loading } = this.props.verificationEmail.toJS();
+
+    const userFields = settings.userFields || [];
+    const languageDictionary = this.props.languageDictionary.get('record').toJS();
+
+    const messageFormat = languageDictionary.resendVerificationEmailMessage ||
+      'Do you really want to resend verification email to {username}?';
+
+    const message = getDialogMessage(messageFormat, 'username',
+      getName(user, userFields, languageDictionary));
 
     return (
-      <Confirm title="Resend Verification Email?" show={requesting} loading={loading} onCancel={cancelResendVerificationEmail} onConfirm={this.onConfirm}>
-        <Error message={error} />
+      <Confirm
+        title={languageDictionary.resendVerificationEmailTitle || "Resend Verification Email?" }
+        show={requesting}
+        loading={loading}
+        confirmMessage={languageDictionary.dialogConfirmText}
+        cancelMessage={languageDictionary.dialogCancelText}
+        onCancel={cancelResendVerificationEmail}
+        closeLabel={languageDictionary.closeButtonText}
+        onConfirm={this.onConfirm}>
+        <Error title={languageDictionary.errorTitle} message={getErrorMessage(languageDictionary, error, settings.errorTranslator)} />
         <p>
-          Do you really want to resend verification email to <strong>{userName}</strong>?
+          {message}
         </p>
       </Confirm>
     );
