@@ -1,63 +1,98 @@
+import _ from 'lodash';
 import React, { Component, PropTypes } from 'react';
 import connectContainer from 'redux-static';
+import { Error, Confirm } from 'auth0-extension-ui';
 
 import { userActions } from '../../../actions';
-import { Error, Confirm } from '../../../components/Dashboard';
+import submitForm from '../../../actions/submitForm';
+import { useDisabledConnectionField, useUsernameField } from '../../../utils/useDefaultFields';
+import { getName, mapValues } from '../../../utils/display';
+import getDialogMessage from './getDialogMessage';
+import UserFieldsForm from '../../../components/Users/UserFieldsForm';
+import getErrorMessage from '../../../utils/getErrorMessage';
 
 export default connectContainer(class extends Component {
   static stateToProps = (state) => ({
-    usernameChange: state.usernameChange
+    usernameChange: state.usernameChange,
+    settings: (state.settings.get('record') && state.settings.get('record').toJS().settings) || {},
+    connections: state.connections,
+    languageDictionary: state.languageDictionary
   });
 
   static actionsToProps = {
+    submitForm,
     ...userActions
-  }
+  };
 
   static propTypes = {
     cancelUsernameChange: PropTypes.func.isRequired,
     changeUsername: PropTypes.func.isRequired,
     usernameChange: PropTypes.object.isRequired
-  }
+  };
 
   shouldComponentUpdate(nextProps) {
-    return nextProps.usernameChange !== this.props.usernameChange;
+    return nextProps.languageDictionary !== this.props.languageDictionary ||
+      nextProps.usernameChange !== this.props.usernameChange;
   }
 
   onConfirm = () => {
-    this.props.changeUsername(this.refs.user.value, this.refs.username.value);
-  }
+    this.props.submitForm('change-username');
+  };
+
+  onSubmit = (formData) => {
+    const languageDictionary = this.props.languageDictionary.get('record').toJS();
+
+    this.props.changeUsername(this.props.usernameChange.toJS().user.user_id, formData, languageDictionary);
+  };
 
   render() {
-    const { cancelUsernameChange } = this.props;
-    const { userId, connection, userNameToChange, userName, error, requesting, loading } = this.props.usernameChange.toJS();
+    const { cancelUsernameChange, connections, settings } = this.props;
+    const { user, connection, error, requesting, loading } = this.props.usernameChange.toJS();
 
     if (!requesting) {
       return null;
     }
 
+    const userFields = settings.userFields || [];
+
+    const languageDictionary = this.props.languageDictionary.get('record').toJS();
+
+    const messageFormat = languageDictionary.changeUsernameMessage ||
+      'Do you really want to change the username for {username}?';
+    const message = getDialogMessage(messageFormat, 'username',
+      getName(user, userFields, languageDictionary));
+
+    const allowedFields = ['username', 'connection'];
+    const initialValues = mapValues(user, allowedFields, userFields, 'edit', languageDictionary);
+    const fields = _.cloneDeep(userFields) || [];
+    useUsernameField(true, fields, connections.get('records').toJS(), connection, initialValues);
+    useDisabledConnectionField(true, fields, connection);
+
+    const filteredFields = _.filter(fields,
+      field => _.includes(allowedFields, field.property));
+
+    const UserFieldsFormInstance = UserFieldsForm('change-username', this.onSubmit);
+
     return (
-      <Confirm title="Change Username?" show={requesting} loading={loading} onCancel={cancelUsernameChange} onConfirm={this.onConfirm}>
-        <Error message={error} />
+      <Confirm
+        title={languageDictionary.changeUsernameTitle || 'Change Username?'}
+        show={requesting}
+        loading={loading}
+        confirmMessage={languageDictionary.dialogConfirmText}
+        cancelMessage={languageDictionary.dialogCancelText}
+        onCancel={cancelUsernameChange}
+        closeLabel={languageDictionary.closeButtonText}
+        onConfirm={this.onConfirm}>
+        <Error title={languageDictionary.errorTitle} message={getErrorMessage(languageDictionary, error, settings.errorTranslator)} />
         <p>
-          Do you really want to change the username for <strong>{userName}</strong>?
+          {message}
         </p>
-        <div className="row">
-          <form className="form-horizontal col-xs-12" style={{ marginTop: '40px' }}>
-            <div className="form-group">
-              <label className="col-xs-2 control-label">Connection</label>
-              <div className="col-xs-9">
-                <input type="text" readOnly="readonly" className="form-control" value={connection} />
-              </div>
-            </div>
-            <div className="form-group">
-              <label className="col-xs-2 control-label">Username</label>
-              <div className="col-xs-9">
-                <input ref="username" type="text" className="form-control" defaultValue={userNameToChange} />
-              </div>
-            </div>
-            <input ref="user" type="hidden" readOnly="readonly" className="form-control" value={userId} />
-          </form>
-        </div>
+        <UserFieldsFormInstance
+          initialValues={initialValues}
+          isEditForm={true}
+          fields={filteredFields}
+          languageDictionary={languageDictionary}
+        />
       </Confirm>
     );
   }
