@@ -6,11 +6,11 @@ import createReducer from '../utils/createReducer';
 const initialState = {
   loading: false,
   error: null,
-  record: { access_level: 0, memberships: [] }
+  record: { role: 0, memberships: [], createMemberships: false }
 };
 
-export const accessLevel = createReducer(fromJS(initialState), {
-  [constants.FETCH_ACCESS_LEVEL]: (state) =>
+export const accessLevel = createReducer(fromJS(initialState), { // eslint-disable-line import/prefer-default-export
+  [constants.FETCH_ACCESS_LEVEL_PENDING]: (state) =>
     state.merge({
       loading: true,
       error: null
@@ -18,7 +18,8 @@ export const accessLevel = createReducer(fromJS(initialState), {
   [constants.FETCH_ACCESS_LEVEL_REJECTED]: (state, action) =>
     state.merge({
       loading: false,
-      error: `An error occured while loading the settings: ${action.errorMessage}`
+      error: action.errorData,
+      record: fromJS({ ...initialState.record, role: 2 })
     }),
   [constants.FETCH_ACCESS_LEVEL_FULFILLED]: (state, action) =>
     state.merge({
