@@ -9,7 +9,7 @@ const initialState = {
   records: []
 };
 
-export const applications = createReducer(fromJS(initialState), {
+export const applications = createReducer(fromJS(initialState), { // eslint-disable-line import/prefer-default-export
   [constants.FETCH_APPLICATIONS_PENDING]: (state) =>
     state.merge({
       loading: true,
@@ -18,7 +18,7 @@ export const applications = createReducer(fromJS(initialState), {
   [constants.FETCH_APPLICATIONS_REJECTED]: (state, action) =>
     state.merge({
       loading: false,
-      error: `An error occured while loading the applications: ${action.errorMessage}`
+      error: action.errorData
     }),
   [constants.FETCH_APPLICATIONS_FULFILLED]: (state, action) =>
     state.merge({
