@@ -3,13 +3,13 @@ import { fromJS } from 'immutable';
 import * as constants from '../constants';
 import createReducer from '../utils/createReducer';
 
-export const scripts = createReducer(fromJS({ }), {
+export const scripts = createReducer(fromJS({ }), { // eslint-disable-line import/prefer-default-export
   [constants.FETCH_SCRIPT_PENDING]: (state, action) =>
     state
-      .setIn([ action.meta.name ], fromJS({ loading: true, error: null, script: null })),
+      .setIn([ action.meta.name ], fromJS({ loading: true, error: null, script: null, token: null })),
   [constants.FETCH_SCRIPT_REJECTED]: (state, action) =>
     state
-      .setIn([ action.meta.name ], fromJS({ loading: false, error: `An error occured while loading the script: ${action.errorMessage}` })),
+      .setIn([ action.meta.name ], fromJS({ loading: false, error: action.errorData })),
   [constants.FETCH_SCRIPT_FULFILLED]: (state, action) =>
     state
       .setIn([ action.meta.name ], fromJS({ loading: false, script: action.payload.data.script })),
@@ -20,7 +20,7 @@ export const scripts = createReducer(fromJS({ }), {
   [constants.UPDATE_SCRIPT_REJECTED]: (state, action) =>
     state
       .setIn([ action.meta.name, 'loading' ], false)
-      .setIn([ action.meta.name, 'error' ], `An error occured while saving the script: ${action.errorMessage}`),
+      .setIn([ action.meta.name, 'error' ], action.errorData),
   [constants.UPDATE_SCRIPT_FULFILLED]: (state, action) =>
     state
       .setIn([ action.meta.name, 'loading' ], false)
