@@ -3,12 +3,12 @@ import { createSelector } from 'reselect';
 const getApps = (state) =>
   state.applications.get('records');
 
-const getConnection = (state, connectionName) =>
-  connectionName && state.connections.get('records').find(conn => conn.get('name') === connectionName);
+const getEnabledClients = (state) =>
+  state.user.get('connection') && state.user.get('connection').get('enabled_clients');
 
 const getAppsForConnection = createSelector(
-  [ getApps, getConnection ],
-  (apps, connection) => apps.filter(app => connection && connection.get('enabled_clients').indexOf(app.get('client_id')) >= 0)
+  [ getApps, getEnabledClients ],
+  (apps, enabledClients) => apps.filter(app => enabledClients && enabledClients.indexOf(app.get('client_id')) >= 0)
 );
 
 export default getAppsForConnection;
