@@ -6,6 +6,13 @@ export default {
       color: 'green'
     }
   },
+  ssa: {
+    event: 'Success Silent Auth',
+    icon: {
+      name: '312',
+      color: 'green'
+    }
+  },
   seacft: {
     event: 'Success Exchange',
     description: 'Authorization Code for Access Token',
@@ -24,6 +31,13 @@ export default {
   },
   f: {
     event: 'Failed Login',
+    icon: {
+      name: '311',
+      color: '#A93F3F'
+    }
+  },
+  fsa: {
+    event: 'Failed Silent Auth',
     icon: {
       name: '311',
       color: '#A93F3F'
@@ -276,6 +290,94 @@ export default {
       name: '311',
       color: '#A93F3F'
     }
+  },
+  seccft: {
+    event: 'Success Exchange (Client Credentials)',
+    icon: {
+      name: '546',
+      color: 'green'
+    }
+  },
+  feccft: {
+    event: 'Failed Exchange (Client Credentials)',
+    icon: {
+      name: '311',
+      color: '#A93F3F'
+    }
+  },
+  limit_mu: {
+    event: 'Blocked IP Address',
+    icon: {
+      name: '311',
+      color: '#A93F3F'
+    }
+  },
+  slo: {
+    event: 'Success Logout',
+    icon: {
+      name: '546',
+      color: 'green'
+    }
+  },
+  flo: {
+    event: 'Failed Logout',
+    icon: {
+      name: '311',
+      color: '#A93F3F'
+    }
+  },
+  sd: {
+    event: 'Success Delegation',
+    icon: {
+      name: '546',
+      color: 'green'
+    }
+  },
+  fd: {
+    event: 'Failed Delegation',
+    icon: {
+      name: '311',
+      color: '#A93F3F'
+    }
+  },
+  gd_start_enroll: {
+    event: 'Enroll started',
+    description: 'Guardian - Start second factor enrollment',
+    icon: {
+      name: '299',
+      color: 'green'
+    }
+  },
+  gd_enrollment_complete: {
+    event: 'Guardian enrollment complete',
+    description: 'Guardian - Enrollment complete',
+    icon: {
+      name: '299',
+      color: 'green'
+    }
+  },
+  gd_auth_succeed: {
+    event: 'OTP Auth suceed',
+    description: 'Guardian - Second factor authentication succeed',
+    icon: {
+      name: 'mfa-login-succeed',
+      color: 'green'
+    }
+  },
+  gd_unenroll: {
+    event: 'Unenroll device account',
+    description: 'Guardian - Enrollment removed',
+    icon: {
+      name: '298',
+      color: 'green'
+    }
+  },
+  sui: {
+    event: 'Users import',
+    description: 'Users import finished',
+    icon: {
+      name: '299',
+      color: 'green'
+    }
   }
-}
-;
+};
