@@ -1,3 +1,3 @@
-export getUserAccessLevel from './getUserAccessLevel';
-export hasAccessLevel from './hasAccessLevel';
 export verifyUserAccess from './verifyUserAccess';
+export requireScope from './requireScope';
+export errorHandler from './errorHandler';
