@@ -3,6 +3,7 @@ import safeEval from 'safe-eval';
 import memoizer from 'lru-memoizer';
 import { ArgumentError } from 'auth0-extension-tools';
 
+import logger from './logger';
 import parseScriptError from './errors/parseScriptError';
 
 export default class ScriptManager {
@@ -11,6 +12,8 @@ export default class ScriptManager {
       throw new ArgumentError('Must provide a storage object.');
     }
 
+    this.log = logger.debug.bind(logger);
+    this.cache = { };
     this.storage = storage;
     this.getCached = Promise.promisify(
       memoizer({
@@ -22,7 +25,7 @@ export default class ScriptManager {
             })
             .catch(callback);
         },
-        hash: (name) => name,
+        hash: name => name,
         max: 100,
         maxAge: cacheAge
       })
@@ -31,7 +34,7 @@ export default class ScriptManager {
 
   get(name) {
     return this.storage.read()
-      .then(data => {
+      .then((data) => {
         if (!data || !data.scripts) {
           return null;
         }
@@ -42,7 +45,7 @@ export default class ScriptManager {
 
   save(name, script) {
     return this.storage.read()
-      .then(data => {
+      .then((data) => {
         if (!data.scripts) {
           data.scripts = {};
         }
@@ -53,24 +56,61 @@ export default class ScriptManager {
       .then(data => this.storage.write(data));
   }
 
+  readCustomData() {
+    return this.storage.read()
+      .then(data => data.customData || { });
+  }
+
+  writeCustomData(customData) {
+    return this.storage.read()
+      .then((data) => {
+        data.customData = customData;
+        return data;
+      })
+      .then(data => this.storage.write(data));
+  }
+
+  createContext(ctx) {
+    return {
+      log: this.log,
+      global: this.cache,
+      read: this.readCustomData.bind(this),
+      write: this.writeCustomData.bind(this),
+      ...ctx
+    };
+  }
+
+  dynamicRequire(module) {
+    try {
+      return __non_webpack_require__(module);
+    } catch(e) {
+      // silently ignore and then try require
+    }
+
+    return require(module);
+  }
+
   execute(name, ctx) {
     return this.getCached(name)
-      .then(script => {
+      .then((script) => {
         if (!script) {
           return null;
         }
 
+        logger.debug(`Executing Delegated Admin hook: ${name}`);
         return new Promise((resolve, reject) => {
           try {
-            const func = safeEval(script);
-            func(ctx, (err, res) => {
+            const func = safeEval(script, { require: this.dynamicRequire });
+            func(this.createContext(ctx), (err, res) => {
               if (err) {
+                logger.error(`Failed to execute Delegated Admin hook (${name}): "${err}"`);
                 reject(parseScriptError(err, name));
               } else {
                 resolve(res);
               }
             });
           } catch (err) {
+            logger.error(`Failed to compile Delegated Admin hook (${name}): "${err}"`);
             reject(parseScriptError(err, name));
           }
         });
