@@ -1,13 +1,15 @@
 import _ from 'lodash';
 import { Router } from 'express';
 
+import multipartRequest from '../lib/multipartRequest';
+
 export default () => {
   const api = Router();
   api.get('/', (req, res, next) => {
-    req.auth0.clients.getAll({ fields: 'client_id,name,callbacks,global' })
+    multipartRequest(req.auth0, 'clients', { is_global: false, fields: 'client_id,name,callbacks,app_type' })
       .then(clients => _.chain(clients)
-        .filter({ global: false })
-        .sortBy((client) => client.name.toLowerCase())
+        .filter(client => client.app_type === 'spa' || client.app_type === 'native' || client.app_type === 'regular_web')
+        .sortBy(client => client.name.toLowerCase())
         .value()
       )
       .then(clients => res.json(clients))
