@@ -1,17 +1,18 @@
 import { Router as router } from 'express';
+import { middlewares } from 'auth0-extension-express-tools';
 
 import config from '../lib/config';
 import logger from '../lib/logger';
-import { middlewares } from 'auth0-extension-express-tools';
+
 
 export default () => {
   const hookValidator = middlewares
-    .validateHookToken(config('AUTH0_DOMAIN'), config('WT_URL'), config('EXTENSION_SECRET'));
+    .validateHookToken(config('AUTH0_ISSUER_DOMAIN'), config('WT_URL'), config('EXTENSION_SECRET'));
 
   const hooks = router();
   hooks.use('/on-uninstall', hookValidator('/.extensions/on-uninstall'));
   hooks.use(middlewares.managementApiClient({
-    domain: config('AUTH0_DOMAIN'),
+    domain: config('AUTH0_ISSUER_DOMAIN'),
     clientId: config('AUTH0_CLIENT_ID'),
     clientSecret: config('AUTH0_CLIENT_SECRET')
   }));
