@@ -1,11 +1,105 @@
+import _ from 'lodash';
 import auth0 from 'auth0';
 import request from 'request';
 import Promise from 'bluebird';
 import { Router } from 'express';
-import { managementApi, ArgumentError, ValidationError } from 'auth0-extension-tools';
+import { ArgumentError, ValidationError } from 'auth0-extension-tools';
 
 import config from '../lib/config';
+import logger from '../lib/logger';
 import { verifyUserAccess } from '../lib/middlewares';
+import { removeGuardian, requestGuardianEnrollments } from '../lib/removeGuardian';
+import requestUserBlocks from '../lib/userBlocks';
+import getApiToken from '../lib/getApiToken';
+import getConnectionIdByName from '../lib/getConnectionIdByName';
+
+const isValidField = (type, onlyTheseFields, field) =>
+  ((onlyTheseFields && _.includes(onlyTheseFields, field.property)) || (!onlyTheseFields && field[type] !== false));
+
+const checkCustomFieldValidation = (req, context, isEditRequest, onlyTheseFields) => {
+  /* Exit early if no custom fields */
+  if (!context.userFields) return context.payload;
+
+  const requiredErrorText = (req.query && req.query.requiredErrorText) || 'required';
+
+  const type = isEditRequest ? 'edit' : 'create';
+
+  /* Loop through valid fields and apply validation function */
+  const ignoredFields = _.map(_.filter(context.userFields, field => !isValidField(type, onlyTheseFields, field)), 'property');
+  const fieldsToValidate = _.filter(context.userFields, field => !_.includes(ignoredFields, field.property) && _.isObject(field[type]) && (field[type].required || field[type].validationFunction));
+
+  const errorList = {};
+  fieldsToValidate.forEach((field) => {
+    const value = _.get(context.payload, field.property);
+    const errorKey = field.label || field.property;
+    if (!value || value.length === 0) {
+      if (field[type].required) {
+        errorList[errorKey] = [ requiredErrorText || 'required' ];
+      }
+      return;
+    }
+
+    if (field[type].validationFunction) {
+      try {
+        const validationFunction = eval(`(${field[type].validationFunction})`);
+        if (!_.isFunction(validationFunction)) {
+          logger.warn(`warning: skipping invalid validation function: ${field[type].validationFunction}, because: it is not a function`);
+        } else {
+          const error = validationFunction(value, context.payload);
+          if (error) {
+            errorList[errorKey] = error;
+            return;
+          }
+        }
+      } catch (e) {
+        logger.warn(`warning: skipping invalid validation function: ${field[type].validationFunction}, because: `, e.message);
+      }
+    }
+
+    if (field[type].options) {
+      const optionValue = _.isObject(value) ? value.value : value;
+      const options = _.map(field[type].options, option => (_.isObject(option) ? option.value : option));
+      if (options.indexOf(optionValue) < 0) {
+        errorList[errorKey] = `${optionValue} is not an allowed option`;
+        return;
+      }
+    }
+  });
+
+  if (Object.keys(errorList).length > 0) throw new ValidationError(_.map(errorList, (value, index) => `${index}: ${value}`).join('\n'));
+
+  /* remove fields from payload that have [type] false */
+  if (onlyTheseFields) {
+    const oldPayload = _.cloneDeep(context.payload);
+    context.payload = {};
+    onlyTheseFields.forEach(key => {
+      _.set(context.payload, key, _.get(oldPayload, key));
+    });
+  }
+
+  return _.omit(context.payload, ignoredFields);
+};
+
+const executeWriteHook = (req, scriptManager, userFields, onlyTheseFields) => {
+  const user = req.targetUser;
+  const context = {
+    method: 'update',
+    request: {
+      user: req.user,
+      originalUser: user
+    },
+    payload: req.body,
+    userFields
+  };
+
+  try {
+    context.payload = checkCustomFieldValidation(req, context, true, onlyTheseFields);
+  } catch (e) {
+    return Promise.reject(e);
+  }
+
+  return scriptManager.execute('create', context);
+};
 
 export default (storage, scriptManager) => {
   const api = Router();
@@ -14,30 +108,62 @@ export default (storage, scriptManager) => {
    * Create user.
    */
   api.post('/', (req, res, next) => {
-    if (!req.body.email || req.body.email.length === 0) {
-      return next(new ValidationError('The email address is required.'));
-    }
-    if (req.body.password !== req.body.repeatPassword) {
-      return next(new ValidationError('The passwords do not match.'));
-    }
-
-    const createContext = {
+    const settingsContext = {
       request: {
         user: req.user
       },
-      payload: {
-        email: req.body.email,
-        username: req.body.username,
-        connection: req.body.connection,
-        memberships: req.body.memberships,
-        password: req.body.password
-      }
+      locale: req.headers['dae-locale']
     };
 
-    return scriptManager.execute('create', createContext)
-      .then(result => req.auth0.users.create(result || createContext.payload))
-      .then(() => res.status(201).send())
-      .catch(next);
+    scriptManager.execute('settings', settingsContext)
+      .then((settings) => {
+        const userFields = settings && settings.userFields;
+        const createContext = {
+          method: 'create',
+          request: {
+            user: req.user
+          },
+          payload: req.body,
+          defaultPayload: {
+            email: req.body.email,
+            username: req.body.username,
+            password: req.body.password,
+            connection: req.body.connection,
+            app_metadata: (req.body.memberships && req.body.memberships.length && { memberships: req.body.memberships }) || {}
+          },
+          userFields
+        };
+
+        const repeatPasswordField = _.find(userFields, { property: 'repeatPassword' });
+        if (!repeatPasswordField) {
+          if (req.body.repeatPassword !== req.body.password) {
+            return next(new ValidationError('The passwords do not match.'));
+          }
+        }
+
+        try {
+          createContext.payload = checkCustomFieldValidation(req, createContext);
+        } catch (e) {
+          return next(e);
+        }
+
+        // Remove repeatPassword before trying to update the user, it has served its purpose
+        delete createContext.payload.repeatPassword;
+
+        return scriptManager.execute('create', createContext)
+          .then((payload) => {
+            // need to preserve the original behavior for null create scripts
+            payload = payload || createContext.defaultPayload;
+            if (!payload.email || payload.email.length === 0) {
+              throw new ValidationError('The email address is required.');
+            }
+
+            return payload;
+          })
+          .then(payload => req.auth0.users.create(payload))
+          .then(() => res.status(201).send())
+          .catch(next);
+      });
   });
 
   /*
@@ -49,55 +175,178 @@ export default (storage, scriptManager) => {
         user: req.user
       },
       payload: {
-        search: req.query.search
+        search: req.query.search,
+        filterBy: req.query.filterBy
       }
     };
 
+    let searchQuery = req.query.search;
+    if (req.query.filterBy && req.query.filterBy.length > 0) {
+      searchQuery = `${req.query.filterBy}:"${req.query.search}"`;
+    }
+    const sort = req.query.sortProperty && req.query.sortOrder
+      ? `${req.query.sortProperty}:${req.query.sortOrder}`
+      : 'last_login:-1';
+
     scriptManager.execute('filter', filterContext)
-      .then(filter => {
+      .then((filter) => {
+        const searchEngine = filter && filter.searchEngine;
+        const defaultEngine = (config('AUTH0_RTA').replace('https://', '') !== 'auth0.auth0.com') ? 'v2' : 'v3';
+        const filterQuery = (filter && typeof filter.query !== 'undefined') ? filter.query : filter;
         const options = {
-          sort: 'last_login:-1',
-          q: (req.query.search && filter) ? `(${req.query.search}) AND ${filter}` : req.query.search || filter,
-          per_page: req.query.per_page || 100,
+          sort,
+          q: (searchQuery && filterQuery) ? `(${searchQuery}) AND ${filterQuery}` : searchQuery || filterQuery,
+          per_page: req.query.per_page || 10,
           page: req.query.page || 0,
           include_totals: true,
-          fields: 'user_id,name,email,identities,picture,last_login,logins_count,multifactor,blocked,app_metadata',
-          search_engine: 'v2'
+          fields: 'user_id,username,name,email,identities,picture,last_login,logins_count,multifactor,blocked,app_metadata,user_metadata',
+          search_engine: searchEngine || (config('SEARCH_ENGINE') !== 'default' && config('SEARCH_ENGINE')) || defaultEngine
         };
 
         return req.auth0.users.getAll(options);
       })
       .then(data =>
-        Promise.map(data.users, (user) =>
-          scriptManager.execute('access', { request: { user: req.user }, payload: { user, action: 'read:user' } }))
+        Promise.map(data.users, (user, index) =>
+          scriptManager.execute('access', { request: { user: req.user }, payload: { user, action: 'read:user' } })
+            .then((parsedUser) => {
+              data.users[index] = parsedUser || user;
+            }))
           .then(() => data))
       .then(users => res.json(users))
-      .catch(next);
+      .catch(err => next(err));
   });
 
   /*
    * Get a single user.
    */
   api.get('/:id', verifyUserAccess('read:user', scriptManager), (req, res, next) => {
-    req.auth0.users.get({ id: req.params.id })
-      .then(user => {
-        const membershipContext = {
-          request: { user }
-        };
+    const user = req.targetUser;
+    const membershipContext = {
+      request: {
+        user: req.user
+      },
+      payload: {
+        user
+      }
+    };
 
-        return scriptManager.execute('memberships', membershipContext)
-          .then(memberships => ({ user, memberships }));
+    return scriptManager.execute('memberships', membershipContext)
+      .then((result) => {
+        if (result && Array.isArray(result)) {
+          return {
+            user,
+            memberships: result
+          };
+        }
+
+        if (result && result.memberships) {
+          return {
+            user,
+            memberships: result.memberships
+          };
+        }
+
+        return {
+          user,
+          memberships: []
+        };
       })
-      .then(data => res.json(data))
-      .catch(next);
+      .then((data) => {
+        if (!data.user.identities) {
+          data.connection = {};
+          return data;
+        }
+
+        const identities = data.user.identities.filter(identity => identity.provider === 'auth0');
+        const name = identities[0] && identities[0].connection;
+
+        if (!name) {
+          data.connection = {};
+          return data;
+        }
+
+        return getConnectionIdByName(req.auth0, name)
+          .then((connectionId) => {
+            if (connectionId) {
+              return req.auth0.connections.get({ id: connectionId, fields: 'enabled_clients' });
+            }
+
+            return {};
+          })
+          .then((connection) => {
+            data.connection = {
+              enabled_clients: connection.enabled_clients
+            };
+
+            return data;
+          });
+      })
+      .then(data =>
+        getApiToken(req)
+          .then(accessToken =>
+            requestUserBlocks(accessToken, req.params.id)
+              .then((blockedFor) => {
+                if (blockedFor) data.user.blocked_for = blockedFor;
+                return accessToken;
+              }))
+          .then((accessToken) => {
+            if (data.user.multifactor && data.user.multifactor.indexOf('guardian') >= 0) {
+              requestGuardianEnrollments(accessToken, req.params.id)
+                .then((enrollments) => {
+                  if (!enrollments || !enrollments.length) {
+                    data.user.multifactor = data.user.multifactor.filter(item => item !== 'guardian');
+                    data.user.multifactor = data.user.multifactor.length ? data.user.multifactor : null;
+                  }
+
+                  return res.json(data);
+                });
+            }
+
+            return res.json(data);
+          })
+      )
+      .catch((err) => {
+        logger.error('Failed to get user because: ', err);
+        next(err);
+      });
   });
 
   /*
    * Deleta a user.
    */
   api.delete('/:id', verifyUserAccess('delete:user', scriptManager), (req, res, next) => {
-    req.auth0.users.delete({ id: req.params.id })
+    if (req.user.sub === req.params.id) {
+      return next(new ValidationError('You cannot delete yourself.'));
+    }
+
+    return req.auth0.users.delete({ id: req.params.id })
       .then(() => res.sendStatus(204))
+      .catch(next);
+  });
+
+  /*
+   * Patch a user.
+   */
+  api.patch('/:id', verifyUserAccess('change:profile', scriptManager), (req, res, next) => {
+    const settingsContext = {
+      request: {
+        user: req.user
+      },
+      locale: req.headers['dae-locale']
+    };
+
+    scriptManager.execute('settings', settingsContext)
+      .then((settings) => {
+        const defaultFields = [ 'username', 'email', 'password', 'repeatPassword', 'connection' ];
+        const allowedFields = _.map(
+          _.filter(settings.userFields,
+            field => field.edit &&
+              !_.includes(defaultFields, field.property)
+          ), 'property');
+        return executeWriteHook(req, scriptManager, settings.userFields, allowedFields);
+      })
+      .then(payload => req.auth0.users.update({ id: req.params.id }, payload))
+      .then(() => res.status(204).send())
       .catch(next);
   });
 
@@ -106,13 +355,13 @@ export default (storage, scriptManager) => {
    */
   api.post('/:id/password-reset', verifyUserAccess('reset:password', scriptManager), (req, res, next) => {
     const client = new auth0.AuthenticationClient({
-      domain: config('AUTH0_DOMAIN'),
+      domain: config('AUTH0_ISSUER_DOMAIN'),
       clientId: config('AUTH0_CLIENT_ID')
     });
 
-    req.auth0.users.get({ id: req.params.id, fields: 'email' })
-      .then(user => ({ email: user.email, connection: req.body.connection, client_id: req.body.clientId }))
-      .then(data => client.requestChangePasswordEmail(data))
+    const user = req.targetUser;
+    const data = { email: user.email, connection: req.body.connection, client_id: req.body.clientId };
+    return client.requestChangePasswordEmail(data)
       .then(() => res.sendStatus(204))
       .catch(next);
   });
@@ -121,35 +370,120 @@ export default (storage, scriptManager) => {
    * Change the password of a user.
    */
   api.put('/:id/change-password', verifyUserAccess('change:password', scriptManager), (req, res, next) => {
-    if (req.body.password !== req.body.confirmPassword) {
+    if (req.body.password !== req.body.repeatPassword) {
       return next(new ArgumentError('Passwords don\'t match'));
     }
 
-    return req.auth0.users.update({ id: req.params.id }, {
-      password: req.body.password,
-      connection: req.body.connection,
-      verify_password: false
-    })
-      .then(() => res.sendStatus(204))
+    const settingsContext = {
+      request: {
+        user: req.user
+      },
+      locale: req.headers['dae-locale']
+    };
+
+    return scriptManager.execute('settings', settingsContext)
+      .then((settings) => {
+        // If userFields is specified in the settings hook, then call the write hook and pass the userFields.
+        if (settings && settings.userFields) {
+          return executeWriteHook(req, scriptManager, settings.userFields, [ 'password', 'repeatPassword' ])
+            .then((payload) => {
+              if (!payload.password) {
+                throw new ValidationError('The password is required.');
+              }
+
+              // Allow app_metadata in case someone needs to set a field in app_metadata to store a flag associated
+              // with the change
+              payload = _.pick(payload, [ 'password', 'connection', 'verify_password', 'app_metadata' ]);
+
+              const payloadFinal = _.defaults(payload, {
+                connection: req.body.connection,
+                verify_password: false
+              });
+
+              return req.auth0.users.update({ id: req.params.id }, payloadFinal)
+                .then(() => res.sendStatus(204))
+                .catch(next);
+            });
+        }
+
+        return req.auth0.users.update({ id: req.params.id }, {
+          password: req.body.password,
+          connection: req.body.connection,
+          verify_password: false
+        })
+          .then(() => res.sendStatus(204))
+          .catch(next);
+      })
       .catch(next);
   });
-
 
   /*
    * Change the username of a user.
    */
-  api.put('/:id/change-username', verifyUserAccess('change:username', scriptManager), (req, res, next) =>
-    req.auth0.users.update({ id: req.params.id }, { username: req.body.username })
-      .then(() => res.sendStatus(204))
-      .catch(next));
+  api.put('/:id/change-username', verifyUserAccess('change:username', scriptManager), (req, res, next) => {
+    const settingsContext = {
+      request: {
+        user: req.user
+      },
+      locale: req.headers['dae-locale']
+    };
+
+    scriptManager.execute('settings', settingsContext)
+      .then((settings) => {
+        // If userFields is specified in the settings hook, then call the write hook and pass the userFields.
+        if (settings && settings.userFields) {
+          executeWriteHook(req, scriptManager, settings.userFields, [ 'username' ])
+            .then((payload) => {
+              if (!payload.username) {
+                throw new ValidationError('The username is required.');
+              }
+
+              return req.auth0.users.update({ id: req.params.id }, payload);
+            })
+            .then(() => res.status(204).send())
+            .catch(next);
+        } else {
+          req.auth0.users.update({ id: req.params.id }, { username: req.body.username })
+            .then(() => res.sendStatus(204))
+            .catch(next);
+        }
+      })
+      .catch(next);
+  });
 
   /*
    * Change the email of a user.
    */
-  api.put('/:id/change-email', verifyUserAccess('change:email', scriptManager), (req, res, next) =>
-    req.auth0.users.update({ id: req.params.id }, { email: req.body.email })
-      .then(() => res.sendStatus(204))
-      .catch(next));
+  api.put('/:id/change-email', verifyUserAccess('change:email', scriptManager), (req, res, next) => {
+    const settingsContext = {
+      request: {
+        user: req.user
+      },
+      locale: req.headers['dae-locale']
+    };
+
+    scriptManager.execute('settings', settingsContext)
+      .then((settings) => {
+        // If userFields is specified in the settings hook, then call the write hook and pass the userFields.
+        if (settings && settings.userFields) {
+          executeWriteHook(req, scriptManager, settings.userFields, [ 'email' ])
+            .then((payload) => {
+              if (!payload.email) {
+                throw new ValidationError('The email is required.');
+              }
+
+              return req.auth0.users.update({ id: req.params.id }, payload);
+            })
+            .then(() => res.status(204).send())
+            .catch(next);
+        } else {
+          req.auth0.users.update({ id: req.params.id }, { email: req.body.email })
+            .then(() => res.sendStatus(204))
+            .catch(next);
+        }
+      })
+      .catch(next);
+  });
 
   /*
    * Get all devices for the user.
@@ -164,10 +498,10 @@ export default (storage, scriptManager) => {
    * Get all logs for a user.
    */
   api.get('/:id/logs', verifyUserAccess('read:logs', scriptManager), (req, res, next) => {
-    managementApi.getAccessTokenCached(config('AUTH0_DOMAIN'), config('AUTH0_CLIENT_ID'), config('AUTH0_CLIENT_SECRET'))
-      .then(accessToken => {
+    getApiToken(req)
+      .then((accessToken) => {
         const options = {
-          uri: `https://${config('AUTH0_DOMAIN')}/api/v2/users/${encodeURIComponent(req.params.id)}/logs`,
+          uri: `https://${config('AUTH0_ISSUER_DOMAIN')}/api/v2/users/${encodeURIComponent(req.params.id)}/logs`,
           qs: {
             include_totals: true
           },
@@ -177,19 +511,18 @@ export default (storage, scriptManager) => {
           json: true
         };
 
-        request.get(options, (err, response, body) => {
+        return request.get(options, (err, response, body) => {
           if (err) {
             return next(err);
           }
 
           if (response.statusCode < 200 || response.statusCode >= 300) {
-            return next(new Error(body && (body.error || body.message || body.code) || `Request Error: ${response.statusCode}`));
+            logger.error('Log response failed: ', response.headers);
+            return next(new Error((body && (body.error || body.message || body.code)) || `Request Error: ${response.statusCode}`));
           }
 
           return res.json(body);
         });
-
-        return request.get(options);
       })
       .catch(next);
   });
@@ -197,8 +530,16 @@ export default (storage, scriptManager) => {
   /*
    * Remove MFA for the user.
    */
-  api.delete('/:id/multifactor/:provider', verifyUserAccess('remove:multifactor-provider', scriptManager), (req, res, next) => {
-    req.auth0.users.deleteMultifactorProvider({ id: req.params.id, provider: req.params.provider })
+  api.delete('/:id/multifactor', verifyUserAccess('remove:multifactor-provider', scriptManager), (req, res, next) => {
+    const providers = req.body.provider || [];
+
+    Promise.map(providers, (provider) => {
+      if (provider !== 'guardian') {
+        return req.auth0.users.deleteMultifactorProvider({ id: req.params.id, provider });
+      }
+
+      return getApiToken(req).then(accessToken => removeGuardian(accessToken, req.params.id));
+    })
       .then(() => res.sendStatus(204))
       .catch(next);
   });
